src/dashboard/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
nats-py>=2.10.0
pydantic>=2.0.0
python-multipart>=0.0.9
